requirements.txt
mcp
httpx
pillow
aiofiles
asyncio-mqtt
pydantic
uvloop
fastapi
uvicorn[standard]
python-dotenv